+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Enabled supported box chars for legacy Windows, and introduce `safe_box` flag
 - Disable hyperlinks on legacy Windows
-- Constructors for Rule and Panel now have keyword only arguments (reson for major version bump)
+- Constructors for Rule and Panel now have keyword only arguments (reason for major version bump)
 
 ### Fixed