+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.1] - 2020-06-30
+
+### Added
+
+- Added box.ASCII2
+- Added markup argument to logging extra
+
+### Changed
+
+- Setting a non-None width now implies expand=True
+
 ## [3.0.0] - 2020-06-28
 
 ### Changed